requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
pystac==1.9.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
